+++ src/admin/src/component/Manage/index.js
@@ -2,21 +2,18 @@ import { Switch, Route } from 'react-router-dom';
 
 import React from 'react';
 
-import Layout from '../Layout';
-
-import Schema from '../Schema';
-import Settings from '../Settings';
-import Dashboard from '../Dashboard';
+import DynaForm from '../DynaForm';
+import DynaView from '../DynaView';
+import DynaTable from '../DynaTable';
 
 const Component = (props) => {
     return (
-        <Layout>
-            <Switch>
-                <Route path={`/settings`} component={Settings} />
-                <Route path={`/manage/:schema`} component={Schema} />
-                <Route path={`/`} component={Dashboard} />
-            </Switch>
-        </Layout>
+        <Switch>
+            <Route path="/create" component={DynaForm} />
+            <Route path="/update/:id" component={DynaForm} />
+            <Route path="/preview/:id" component={DynaView} />
+            <Route path="/" component={DynaTable} />
+        </Switch>
     )
 };
